+++ database/migrations/2025_07_21_144443_create_vendor_bills_table.php
@@ -23,8 +23,8 @@ return new class extends Migration
             $table->date('accounting_date');
             $table->date('due_date')->nullable();
             $table->string('status')->default(VendorBill::TYPE_DRAFT)->index(); // 'draft', 'posted', 'paid'
-            $table->decimal('total_amount', 15, 2);
-            $table->decimal('total_tax', 15, 2);
+            $table->unsignedBigInteger('total_amount');
+            $table->unsignedBigInteger('total_tax');
             $table->timestamp('posted_at')->nullable();
             $table->json('reset_to_draft_log')->nullable();
             $table->timestamps();
